Ticket Price calculator
This program determines how much a ticket will cost for each user. The user is able to enter how many tickets they need.  They are also able to add a few features to their purchase for an addtional cost. The program determines the user input and calculates the price for them.