3D printing project
This program is a calculator that takes user input to determine what the SUBTOTAL, SHIPPING COST, TAX, and GRAND TOTAL will be.

In this program we see that 3 types of input method are used.  A selection box is used, a user input box, and radial buttons are used as well.