Commission project
This program displays how much commission an employee has earned. The user types in the employee ID and the program displays that employee's commission.

The program is connected to a .CSV file and locates the employee ID on that file. It then grabs the sales commission from the .CSV file and displays it within the program.